Fix accel read, and database write
Fixed accel read to read signed values properly
Fixed the database write by having a placeholder array with the maximum number of users for now.
Also added a method to display the current accel value when "accelTimer" is sent as a command in console mode.  Console mode can be entered with a button press now