Added Inactivity checking
The current method is comparing an average acceleration value on all 3 axes over a specific time to the current value.  If the difference is less than a tolerance the vehicle is considered inactive.  This may have scaling issues to a larger time period due to the growth of required memory.